optimization of the scene adding
I was adding the scenes before creating the config of the game with await. This was making the loading before the game instance too long proportional to the number of scenes.
So now I am only adding the boot scene there and the others I am adding in the boot scene and the preload scenes.